Stop dispatching one job per workspace
