+++ packages/twenty-server/src/modules/workflow/workflow-runner/workflow-run-queue/cron/jobs/workflow-handle-staled-runs.job.ts
@@ -1,19 +1,19 @@
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { WorkspaceActivationStatus } from 'twenty-shared/workspace';
-import { Repository } from 'typeorm';
+import { IsNull, LessThan, Or, Repository } from 'typeorm';
 
 import { SentryCronMonitor } from 'src/engine/core-modules/cron/sentry-cron-monitor.decorator';
-import { InjectMessageQueue } from 'src/engine/core-modules/message-queue/decorators/message-queue.decorator';
 import { Process } from 'src/engine/core-modules/message-queue/decorators/process.decorator';
 import { Processor } from 'src/engine/core-modules/message-queue/decorators/processor.decorator';
 import { MessageQueue } from 'src/engine/core-modules/message-queue/message-queue.constants';
-import { MessageQueueService } from 'src/engine/core-modules/message-queue/services/message-queue.service';
 import { Workspace } from 'src/engine/core-modules/workspace/workspace.entity';
+import { TwentyORMGlobalManager } from 'src/engine/twenty-orm/twenty-orm-global.manager';
 import {
-  WorkflowHandleStaledRunsPerWorkspaceJob,
-  type WorkflowHandleStaledRunsPerWorkspaceJobData,
-} from 'src/modules/workflow/workflow-runner/workflow-run-queue/cron/jobs/workflow-handle-staled-runs-per-workspace.job';
+  WorkflowRunStatus,
+  WorkflowRunWorkspaceEntity,
+} from 'src/modules/workflow/common/standard-objects/workflow-run.workspace-entity';
+import { WorkflowRunQueueWorkspaceService } from 'src/modules/workflow/workflow-runner/workflow-run-queue/workspace-services/workflow-run-queue.workspace-service';
 
 export const WORKFLOW_HANDLE_STALED_RUNS_CRON_PATTERN = '0 * * * *';
 
@@ -22,8 +22,8 @@ export class WorkflowHandleStaledRunsJob {
   constructor(
     @InjectRepository(Workspace)
     private readonly workspaceRepository: Repository<Workspace>,
-    @InjectMessageQueue(MessageQueue.workflowQueue)
-    private readonly messageQueueService: MessageQueueService,
+    private readonly workflowRunQueueWorkspaceService: WorkflowRunQueueWorkspaceService,
+    private readonly twentyORMGlobalManager: TwentyORMGlobalManager,
   ) {}
 
   @Process(WorkflowHandleStaledRunsJob.name)
@@ -39,12 +39,45 @@ export class WorkflowHandleStaledRunsJob {
     });
 
     for (const activeWorkspace of activeWorkspaces) {
-      await this.messageQueueService.add<WorkflowHandleStaledRunsPerWorkspaceJobData>(
-        WorkflowHandleStaledRunsPerWorkspaceJob.name,
-        {
-          workspaceId: activeWorkspace.id,
-        },
-      );
+      try {
+        const workflowRunRepository =
+          await this.twentyORMGlobalManager.getRepositoryForWorkspace(
+            activeWorkspace.id,
+            WorkflowRunWorkspaceEntity,
+            { shouldBypassPermissionChecks: true },
+          );
+
+        const oneHourAgo = new Date(Date.now() - 60 * 60 * 1000);
+
+        const staledWorkflowRuns = await workflowRunRepository.find({
+          where: {
+            status: WorkflowRunStatus.ENQUEUED,
+            enqueuedAt: Or(LessThan(oneHourAgo), IsNull()),
+          },
+        });
+
+        if (staledWorkflowRuns.length <= 0) {
+          continue;
+        }
+
+        await workflowRunRepository.update(
+          staledWorkflowRuns.map((workflowRun) => workflowRun.id),
+          {
+            enqueuedAt: null,
+            status: WorkflowRunStatus.NOT_STARTED,
+          },
+        );
+
+        await this.workflowRunQueueWorkspaceService.recomputeWorkflowRunQueuedCount(
+          activeWorkspace.id,
+        );
+      } catch (error) {
+        console.error(
+          'Failed to handle staled runs for workspace',
+          activeWorkspace.id,
+          error,
+        );
+      }
     }
   }
 }
